#1 - Local build (see desc) + add images for Projects Template (static)
I have two folders on my personal computer, one of them is the folder containing this repo (and the multiple Django projects in it) and the other one is the `django-start` project (i.e. a singular django project that is the same as the one in the `django-start` directory in this repo). I may have missed steps in both projects as I've had two folders, so I'm copying it in here so that we can see missed steps between the two folders
On my desktop pc (not my laptop), I have THIS REPO installed at `users/droid/django-start`, so I can just `cd/django-start/django-start`. On my laptop, I have to `cd/django-start` (which gets me into this singular intro course project to Django) or, to get into this repo, `cd/rp-portfolio/django-start`. Nothing that can't be fixed, however.

Also adding in the images for the Projects [app] template. We can use our own images but just for the sake of following this course we're going to use the demo ones

Every instruction we've followed can be seen in the #1 issue (notion & the teams we've tagged)

Co-Authored-By: Liam Arbuckle <allianceofdroidsla@gmail.com>